Moved logging to own namespace
