+++ src/Lavalink4NET/Logging/EventLogger.cs
@@ -1,5 +1,5 @@
 ﻿/*
- *  File:   ILogger.cs
+ *  File:   EventLogger.cs
  *  Author: Angelo Breuer
  *
  *  The MIT License (MIT)
@@ -25,15 +25,20 @@
  *  THE SOFTWARE.
  */
 
-namespace Lavalink4NET
+namespace Lavalink4NET.Logging
 {
     using System;
 
     /// <summary>
-    ///     An interface for a logger provider.
+    ///     A logger implementation for a logger that logs the messages to an event.
     /// </summary>
-    public interface ILogger
+    public class EventLogger : ILogger
     {
+        /// <summary>
+        ///     An event that is triggered when a message is logged.
+        /// </summary>
+        public event EventHandler<LogMessageEventArgs> LogMessage;
+
         /// <summary>
         ///     Logs a message.
         /// </summary>
@@ -41,6 +46,14 @@ namespace Lavalink4NET
         /// <param name="message">the message to log</param>
         /// <param name="level">the logging level / the severity of the message</param>
         /// <param name="exception">an optional exception that occurred</param>
-        void Log(object source, string message, LogLevel level = LogLevel.Information, Exception exception = null);
+        public void Log(object source, string message, LogLevel level = LogLevel.Information, Exception exception = null)
+            => OnLogMessage(new LogMessageEventArgs(source, message, level, exception));
+
+        /// <summary>
+        ///     Triggers the <see cref="LogMessage"/> event.
+        /// </summary>
+        /// <param name="eventArgs">the event arguments</param>
+        protected virtual void OnLogMessage(LogMessageEventArgs eventArgs)
+            => LogMessage?.Invoke(this, eventArgs);
     }
 }
